refactor(#1358): move outbox utilities to neutral module; eliminate behaviors→triggers import edge
- Move _log_label, outbox_ids, add_activity_to_outbox from
  triggers/_helpers.py to use_cases/_helpers.py (neutral leaf module)
- Delete dead update_participant_rm_state duplicate from triggers/_helpers.py
  (zero callers; live copy already in use_cases/_helpers.py)
- Remove now-unused imports from triggers/_helpers.py: hashlib,
  is_participant_model, RM
- Update 5 production import sites to point to use_cases._helpers:
  behaviors/embargo/nodes/lifecycle.py, behaviors/sender/nodes/actions.py,
  adapters/driven/sync_activity_adapter.py, triggers/report.py,
  received/actor/case_manager_role.py
- Update 3 lazy import sites and 1 mock patch target in tests

The behaviors->triggers module-level import edge is now gone. use_cases/_helpers.py
is a true neutral leaf (ports + stdlib only). The rule 'BT nodes import shared
utilities from use_cases._helpers' is now uniformly true across the codebase.
