+++ vultron/core/behaviors/sender/nodes/actions.py
@@ -23,7 +23,7 @@ from py_trees.common import Status
 from vultron.core.behaviors.helpers import DataLayerAction
 from vultron.core.models.protocols import is_case_model
 from vultron.core.use_cases._helpers import _resolve_case_manager_id
-from vultron.core.use_cases.triggers._helpers import add_activity_to_outbox
+from vultron.core.use_cases._helpers import add_activity_to_outbox
 
 
 class ResolveCaseManagerNode(DataLayerAction):
