+++ vultron/core/use_cases/triggers/_helpers.py
@@ -22,15 +22,12 @@ framework imports allowed here.
 """
 
 import logging
-import hashlib
 from collections.abc import Callable
 
 from vultron.core.models.protocols import (
     CaseModel,
     is_case_model,
-    is_participant_model,
 )
-from vultron.core.states.rm import RM
 from vultron.core.ports.case_persistence import (
     CasePersistence,
     CaseOutboxPersistence,
@@ -39,17 +36,6 @@ from vultron.core.ports.trigger_activity import TriggerActivityPort
 from vultron.errors import VultronNotFoundError, VultronValidationError
 
 logger = logging.getLogger(__name__)
-
-
-def _log_label(uri: str) -> str:
-    """Return a deterministic redacted label for IDs used in log messages.
-
-    Do not log raw actor/activity identifiers (or URI segments) because they
-    may be sensitive.  Instead, emit a short non-reversible hash token that
-    still allows correlation across log lines.
-    """
-    digest = hashlib.sha256(uri.encode("utf-8")).hexdigest()[:12]
-    return f"id:{digest}"
 
 
 def resolve_actor(actor_id: str, dl: CasePersistence):
@@ -74,128 +60,6 @@ def resolve_case(case_id: str, dl: CasePersistence) -> CaseModel:
             f"Expected VulnerabilityCase, got {type(case_raw).__name__}."
         )
     return case_raw
-
-
-def update_participant_rm_state(
-    case_id: str, actor_id: str, new_rm_state: RM, dl: CasePersistence
-) -> bool:
-    """
-    Append a new ParticipantStatus with new_rm_state to the actor's
-    CaseParticipant in the given case and persist the updated participant.
-
-    Handles both inline and string-reference participants.
-
-    Returns ``True`` on success (including idempotent no-op), ``False`` when
-    the case or participant is not found.
-    """
-    case_obj = dl.read(case_id)
-    if not is_case_model(case_obj):
-        logger.warning(
-            "update_participant_rm_state: case '%s' not found or wrong type",
-            case_id,
-        )
-        return False
-
-    for participant_ref in case_obj.case_participants:
-        if isinstance(participant_ref, str):
-            participant_raw = dl.read(participant_ref)
-            if participant_raw is None:
-                continue
-        else:
-            participant_raw = participant_ref
-
-        if not is_participant_model(participant_raw):
-            continue
-
-        participant = participant_raw
-
-        actor_ref = participant.attributed_to
-        p_actor_id = (
-            actor_ref
-            if isinstance(actor_ref, str)
-            else getattr(actor_ref, "id_", str(actor_ref))
-        )
-        if p_actor_id == actor_id:
-            if participant.participant_statuses:
-                latest = participant.participant_statuses[-1]
-                if latest.rm_state == new_rm_state:
-                    logger.info(
-                        "Participant '%s' already in RM state %s in case '%s' "
-                        "(idempotent)",
-                        actor_id,
-                        new_rm_state,
-                        case_id,
-                    )
-                    return True
-            appended = participant.append_rm_state(
-                rm_state=new_rm_state, actor=actor_id, context=case_id
-            )
-            if not appended:
-                logger.warning(
-                    "update_participant_rm_state: RM transition to %s blocked "
-                    "for actor '%s' in case '%s'",
-                    new_rm_state,
-                    actor_id,
-                    case_id,
-                )
-                return False
-            dl.save(participant)
-            logger.info(
-                "Set participant '%s' RM state to %s in case '%s'",
-                actor_id,
-                new_rm_state,
-                case_id,
-            )
-            return True
-
-    logger.warning(
-        "update_participant_rm_state: no CaseParticipant for actor '%s' "
-        "in case '%s'; RM state not updated",
-        actor_id,
-        case_id,
-    )
-    return False
-
-
-def outbox_ids(actor_id: str, dl: CaseOutboxPersistence) -> set[str]:
-    """Return the set of string activity IDs in the actor's outbox queue.
-
-    Uses ``outbox_list_for_actor`` when available (explicit actor scope),
-    otherwise falls back to the actor-scoped ``outbox_list()``.
-
-    Args:
-        actor_id: The actor whose outbox should be queried.
-        dl: The DataLayer to use for persistence.
-
-    Returns:
-        Set of activity IDs queued for delivery.
-    """
-    if hasattr(dl, "outbox_list_for_actor"):
-        items: list[str] = dl.outbox_list_for_actor(actor_id)  # type: ignore[attr-defined]
-        return set(items)
-    return set(dl.outbox_list())
-
-
-def add_activity_to_outbox(
-    actor_id: str, activity_id: str, dl: CaseOutboxPersistence
-) -> None:
-    """Append an activity ID to an actor's outbox and queue it for delivery.
-
-    Uses ``record_outbox_item`` to explicitly enqueue against *actor_id*,
-    bypassing any actor-scope on *dl* itself.  This ensures correct delivery
-    even when *dl* is a shared (unscoped) DataLayer instance.
-
-    Args:
-        actor_id: The actor whose outbox should receive the activity.
-        activity_id: The ID of the activity to queue for delivery.
-        dl: The DataLayer to use for persistence.
-    """
-    dl.record_outbox_item(actor_id, activity_id)
-    logger.debug(
-        "Queued activity '%s' in delivery queue for actor '%s'",
-        _log_label(activity_id),
-        _log_label(actor_id),
-    )
 
 
 def find_embargo_proposal(case_id: str, dl: CasePersistence):
